Restructured and added test migration.
