+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "start": "node index.js",
     "dev": "nodemon index.js",
+    "knex": "knex --knexfile src/db/knexfile.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
